update: move MarkerCollection from markers to pipeline
